Add FDT support (kernel appended and separated DTB).
FDTs are used to encode device/board-specific information such that this
information is kept independent of the driver code.
There are two ways of deploying FDTs in android:
1) It can be appended to the boot image using dbtTool.
2) It can be appended to the kernel.

This commit equips libbootimg with the ability to extract and load FDTs
in both of the aforementioned cases. Loading an FDT is only supported if
the original image (from which it has been extracted before) provided a
FDT and the library loads the image into the bootimage data equivalent
to the status before (appended or separated).

Finding the fdt and generating the needed information is only funtional
if the mask of the blob containing the FDT is given when reading an
bootimage. E.g., for appended FDTs, the kernel mask must be given (see
libbootimg.h for the masks). Hence, providing the masks for the kernel
and the separated DTB ensures that the FDT is found (if one is present).

The main external interface functions are libbootimg_dump_fdt and
libbootimg_load_fdt. Before using these functions, the originating
bootimage must have been read using the aforementioned masks.

In particular, the following changes were done:
* Add a "fdt_info" entry to bootimg structure that references the FDT
  header, information whether it is appended or separated, its start
  address and size.
* Introduce a method that extracts the required information (FDT header,
  blob addresses --> fdt_info from above) and which works for both
  appended and separated DTBs. Functions using this information do not
  need to be aware of the originating location.
* Introduce a search method for finding the "FDT_MAGIC" (start of the
  FDT) in a block of binary data.
* Add a function 'libboot_dump_fdt' that uses the fdt_info information
  about the FDT and writes it to a given file.
* Add a function 'libboot_load_fdt' that loads an FDT blob from a given
  file. The library places the given given blob at the location where
  the FDT has been located formerly when loading the bootimage. Thereby,
  the preceeding information of the FDT is copied to a new memory
  location and the FDT is concernated to it (and padding is performed in
  separate DTB case). Note that separate DTBs are not FDTs, but are
  preceeded by a block of zeros besides the initial QCDT magic. Hence,
  the preceeding data must be also preserved in this case.

Change-Id: Iecd9bac75aef5799c952addc34ab4fd7b1df3d08
Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>